requirement.txt
accelerate==0.22.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.1
bleach==6.0.0
blinker==1.7.0
bot-studio==1.4.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
DateTime==5.4
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
diffusers @ git+https://github.com/huggingface/diffusers@b954c22a446301c644120f41c046a4d4c4553c7a
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.3
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
google-images-download==2.8.0
googletrans==4.0.0rc1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.17.1
hyperframe==5.2.0
idna==2.10
imageio==2.31.3
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.3
jupyter_client==8.0.1
jupyter_core==5.1.5
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.0
kiwisolver==1.4.4
logger==1.4
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.2.4
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.0.10
outcome==1.2.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
proglog==0.1.10
progressbar==2.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.14.0
pygoogle-image==1.0.0
PyJWT==2.8.0
pyotp==2.3.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic-bin==0.4.14
pytz==2023.3.post1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.2
QtPy==2.3.1
referencing==0.32.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.17.1
safetensors==0.3.3
schedule==1.2.0
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
selenium==4.12.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.25
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.12
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
tornado==6.2
tqdm==4.66.1
traitlets==5.8.1
transformers==4.33.1
trio==0.22.2
trio-websocket==0.10.4
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2022.7
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
widgetsnbextension==4.0.7
wsproto==1.2.0
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.16.2
zope.interface==6.2
